% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_analysis.R
\name{DESeq2_function}
\alias{DESeq2_function}
\title{Preform Differential Gene Analysis}
\usage{
DESeq2_function(se, ref_level = "Tconv")
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{ref_level}{sets the reference to a specific cell type (default: "Tconv")}
}
\value{
The results of the Differential Gene Analysis between cell_type

@importFrom SummarizedExperiment assay
@export

@examples
library(SummarizedExperiment)
tissueTreg_filter_DESeq2<- DESeq2_function(example_se)
}
\description{
Preform Differential Gene Analysis
}
