% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_filter_threshold.R
\name{determine_filter_threshold}
\alias{determine_filter_threshold}
\title{The filtering diagnostics evaluate how the pre-filtering threshold (minimum count per group) affects the number of discoveries, enabling users to choose an informed threshold}
\usage{
determine_filter_threshold(
  se_ln,
  count_thresholds = c(0, 1, 5, 10, 20, 50, 100, 200, 500),
  assay_name = "counts",
  ref_level = "Tconv",
  group_var = "cell_type",
  p_threshold = 0.05
)
}
\arguments{
\item{se_ln}{A SummarizedExperiment object}

\item{count_thresholds}{minimum threshold of gene counts (default: 'c(0, 1, 5, 10, 20, 50, 100, 200, 500)')}

\item{assay_name}{Name of assay to use (default: "counts")}

\item{ref_level}{Sets the reference for comparison (default: "Tconv")}

\item{group_var}{metadata for grouping (default: "cell_type")}

\item{p_threshold}{Significance threshold for adjusted p-value(default: 0.05)}
}
\value{
A dataframe summarizing the number of significant genes per threshold
}
\description{
The filtering diagnostics evaluate how the pre-filtering threshold (minimum count per group) affects the number of discoveries, enabling users to choose an informed threshold
}
\examples{
data(example_se)

# Step 1: Evaluate how model preforms using different threshold values
example_se_filtering_assessment<- determine_filter_threshold(
se_ln = example_se,
count_thresholds = c(0, 1, 5, 10, 20, 50, 100, 200, 500),
assay_name = "counts",
ref_level = "Tconv",
group_var = "cell_type",
p_threshold = 0.05)

# Step 2: Pick the best min count threshold and set \min_count_per_group` (default: 10)`

}
