% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_outputs}
\alias{export_outputs}
\title{Export differential expression results, summary counts, and filtering diagnostics to TSV files}
\usage{
export_outputs(
  res_df,
  summary_df,
  filtering_diag,
  volcano,
  output_dir = file.path(tempdir(), "de_output")
)
}
\arguments{
\item{res_df}{Dataframe of log2fold-changed shrinkage applied to ds}

\item{summary_df}{A data frame summarizing the number of non-significant(ns)/up/down genes}

\item{filtering_diag}{A data frame summarizing the number of significant genes per threshold}

\item{volcano}{Your choice of volcano plot}

\item{output_dir}{Directories where files will be written}
}
\value{
TSV files of differential expression results, summary counts, and filtering diagnostics
}
\description{
Export differential expression results, summary counts, and filtering diagnostics to TSV files
}
\examples{
 data(example_se)

# Step 1: Evaluate how model preforms using different threshold values and choose a threshold
example_se_filtering_assessment<- determine_filter_threshold(example_se)

# Step 2; Filter low expression genes
se_filtered<- filter_low_exp_genes(example_se, min_count_per_group = 10)

# Step 3: Run the DESeq2 pipeline to get differential gene expression results
 se_dge<- run_DESeq2(se_filtered)

# Step 4:Run log2_shrinkage on DESeq2 results to improve estimates
se_dge_shrink <- log2_shrinkage(se_dge)

# Step 5:Generate a volcano plot from log2_shrinkage results
example_se_volcano<- generate_volcano(se_dge_shrink)

# Step 6: Generate the regulation summary of the genes
DESeq2_gene_reg_summary<- gene_regulation_summary(se_dge_shrink)

# Step 7: Export final results
example_se_exports<- export_outputs(
res_df = se_dge_shrink,
summary_df = DESeq2_gene_reg_summary,
filtering_diag = example_se_filtering_assessment,
volcano = example_se_volcano,
output_dir = file.path(tempdir(), "de_output") )

}
