% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_exp_gene.R
\name{filter_low_exp_genes}
\alias{filter_low_exp_genes}
\title{Filter genes with low expression}
\usage{
filter_low_exp_genes(se_ln, min_count_per_group = 10, assay_name = "counts")
}
\arguments{
\item{se_ln}{A SummarizedExperiment object}

\item{min_count_per_group}{The minimum threshold of gene count (default: 10)}

\item{assay_name}{Name of assay to use (default: "counts")}
}
\value{
A SummarizedExperiment subset of genes with greater than n counts
}
\description{
Filter genes with low expression
}
\examples{
library(SummarizedExperiment)
data(example_se)

# Step 1; Filter low expression genes
se_filtered <- filter_low_exp_genes(se_ln = example_se, min_count_per_group = 10, assay_name = "counts")

}
