% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{generate_volcano}
\alias{generate_volcano}
\title{Visualize results as a volcano plot}
\usage{
generate_volcano(
  res_df,
  fc_threshold = 0.5,
  xlab = "log2 Fold Change (Treg vs Tconv)",
  set_title = "Volcano Plot - Lymph Node Treg vs Tconv",
  p_threshold = 0.05
)
}
\arguments{
\item{res_df}{Dataframe of log2fold-changed shrinkage applied to ds}

\item{fc_threshold}{(default: 0.5)}

\item{xlab}{The label of the x-axis}

\item{set_title}{The title of the plot}

\item{p_threshold}{THe threshold that defines a gene's significance (default: 0.05)}
}
\value{
volcano pot with significant genes highlighted
}
\description{
Visualize results as a volcano plot
}
\examples{
library(ggplot2)
library(SummarizedExperiment)

tissueTreg_volcano<- generate_volcano(tissueTreg_filter_DESeq2_shrink)

}
