% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{log2_shrinkage}
\alias{log2_shrinkage}
\title{Applies log2fold-change shrinkage for more reliable effect-size estimates}
\usage{
log2_shrinkage(dds, shrinkage = "apeglm")
}
\arguments{
\item{dds}{The results of 'DESeq2_function'}

\item{shrinkage}{The estimator used to assess the glm coefficeints (default: "apeglm")}
}
\value{
Dataframe of log2fold-changed shrinkage applied to ds
}
\description{
Applies log2fold-change shrinkage for more reliable effect-size estimates
}
\examples{
library(SummarizedExperiment)
library(DESeq2)
library(apeglm)

tissueTreg_filter_DESeq2_shrink <- log2_shrinkage(tissueTreg_filter_DESeq2)

}
