% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_analysis.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{Preform Differential Gene Analysis}
\usage{
run_DESeq2(se_ln, group_var = "cell_type", ref_level = "Tconv")
}
\arguments{
\item{se_ln}{A SummarizedExperiment object}

\item{group_var}{determines the column to categorization}

\item{ref_level}{sets the reference to a specific cell type (default: "Tconv")}
}
\value{
The results of the Differential Gene Analysis between cell_type
}
\description{
Preform Differential Gene Analysis
}
\examples{
library(SummarizedExperiment)
library(DESeq2)
data(example_se)

# Step 1; Filter low expression genes
se_filtered<- filter_low_exp_genes(example_se, min_count_per_group = 10)

# Step 2: Run the DESeq2 pipeline to get shrinkage results
se_dge<- run_DESeq2(se_ln = se_filtered, group_var = "cell_type", ref_level = "Tconv")

}
