% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_outputs}
\alias{export_outputs}
\title{Export differential expression results, summary counts, and filtering diagnostics to TSV files}
\usage{
export_outputs(
  res_df,
  summary_df,
  filtering_diag,
  output_dir = file.path(tempdir(), "de_output")
)
}
\arguments{
\item{res_df}{Dataframe of log2fold-changed shrinkage applied to ds}

\item{summary_df}{A data frame summarizing the number of non-significant(ns)/up/down genes}

\item{filtering_diag}{A data frame summarizing the number of significant genes per threshold}

\item{output_dir}{Directories where files will be written}
}
\value{
TSV files of differential expression results, summary counts, and filtering diagnostics
}
\description{
Export differential expression results, summary counts, and filtering diagnostics to TSV files
}
\examples{
export_results<- export_outputs( tissueTreg_filter_DESeq2_shrink, DESeq2_gene_reg_summary,  tissueTreg_filtering_assessment)
}
