% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_genes.R
\name{gene_regulation_summary}
\alias{gene_regulation_summary}
\title{Summarize the number of significantly non-significant(ns)/up/down genes at given thresholds}
\usage{
gene_regulation_summary(res_df, p_threshold = 0.05, fc_threshold = 0.5)
}
\arguments{
\item{res_df}{Dataframe of log2fold-changed shrinkage applied to ds}

\item{p_threshold}{The p-value that determines the number for genes per category(default: 0.05)}

\item{fc_threshold}{(default = 0.5)}
}
\value{
Summary of non-significant(ns)/up/down genes
}
\description{
Summarize the number of significantly non-significant(ns)/up/down genes at given thresholds
}
\examples{
library(SummarizedExperiment)

DESeq2_gene_reg_summary<- gene_regulation_summary(tissueTreg_filter_DESeq2_shrink)

}
