% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{generate_volcano}
\alias{generate_volcano}
\title{Visualize results as a volcano plot}
\usage{
generate_volcano(
  res_df,
  fc_threshold = 0.5,
  xlab = "log2 Fold Change (Treg vs Tconv)",
  set_title = "Volcano Plot - Lymph Node Treg vs Tconv",
  p_threshold = 0.05
)
}
\arguments{
\item{res_df}{Dataframe of log2fold-changed shrinkage applied to ds}

\item{fc_threshold}{fold change threshold(default: 0.5)}

\item{xlab}{The label of the x-axis}

\item{set_title}{The title of the plot}

\item{p_threshold}{The threshold that defines a gene's significance (default: 0.05)}
}
\value{
volcano pot with significant genes highlighted
}
\description{
Visualize results as a volcano plot
}
\examples{
data(example_se)

# Step 1; Filter low expression genes
se_filtered<- filter_low_exp_genes(example_se, min_count_per_group = 10)

# Step 2: Run the DESeq2 pipeline to get differential gene expression results
 se_dge<- run_DESeq2(se_filtered)

# Step 3: Run the log2_shrinkage function on the results of the DESeq2 function to create more reliable estimates
se_dge_shrink <- log2_shrinkage(se_dge)

# Step 4: Create a volcano plot of the gene expression using the results obtained from the log2_shrinkage function
example_se_volcano<- generate_volcano(res_df = se_dge_shrink, fc_threshold =  0.5, xlab = "log2 Fold Change (Treg vs Tconv)", set_title = "Volcano Plot - Lymph Node Treg vs Tconv", p_threshold = 0.05)

}
