% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{log2_shrinkage}
\alias{log2_shrinkage}
\title{Apply log2fold-change shrinkage for more reliable effect-size estimates}
\usage{
log2_shrinkage(dds, shrinkage = "apeglm")
}
\arguments{
\item{dds}{The results of 'DESeq2_function'}

\item{shrinkage}{The estimator used to assess the glm coefficeints (default: "apeglm")}
}
\value{
Dataframe of log2fold-changed shrinkage applied to ds
}
\description{
Apply log2fold-change shrinkage for more reliable effect-size estimates
}
\examples{
data(example_se)

# Step 1; Filter low expression genes
se_filtered<- filter_low_exp_genes(example_se, min_count_per_group = 10)

# Step 2: Run the DESeq2 pipeline to get differential gene expression results
 se_dge<- run_DESeq2(se_filtered)

# Step 3:Run log2_shrinkage on DESeq2 results to improve estimates
se_dge_shrink <- log2_shrinkage(dds = se_dge, shrinkage = "apeglm")

}
